Fix load so that it works again.
However I don't understand the fix. The fix is to use requirejs+jquery in the same library rather than just jquery. Even though I'm not actually using requirejs explicitly. There must be something in the requirejs initializer that's doing the magic. In the future it would be good to track this down.